sports-score-be/app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "motor>=3.7.1",
    "passlib>=1.7.4",
    "pydantic[email]>=2.11.7",
    "python-jose>=3.5.0",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.1",
]
